requirements.txt
# Core numerical & vision
numpy
pandas
Pillow
opencv-python
scipy
scikit-image
scikit-learn
tqdm

# PyTorch & CUDA
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118

# CLIP & Others
git+https://github.com/openai/CLIP.git#egg=clip
timm
einops
transformers>=4.20.0         # for any text‐encoder or tokenizer utilities

# RINE-style contrastive / losses
torchtoolbox

# Wavelet-CLIP extras
PyWavelets
pytorch-wavelets
kornia
filterpy
simplejson

# Augmentations & image I/O
albumentations
imgaug
imageio
imutils

# Logging & visualization
tensorboard

# Pretrained helpers
efficientnet-pytorch
segmentation-models-pytorch
pytorchvideo
loralib

# YAML configs
PyYAML>=6.0